Refactor aisr
